requirements.txt
# idegeo dependencies
django-ckeditor==5.0.0
django-fontawesome==0.3.1
geopandas==0.8.1
pandas==1.1.4
python-sld==1.0.10
csvkit==0.9.2
utm==0.5.0
GeoAlchemy2==0.8.4

# native dependencies
Pillow==8.0.1
lxml==4.6.1
defusedxml==0.6.0
psycopg2==2.8.6
Django==2.2.16

# Other
amqp==5.0.1
pyyaml>=4.2b1
beautifulsoup4==4.9.3
httplib2<0.18.2
hyperlink==20.0.1
idna>=2.5,<2.11
urllib3==1.25.11
Paver==1.3.4
python-slugify==4.0.1
decorator==4.4.2
celery==5.0.1
kombu==5.0.2
vine==5.0.0
boto3==1.16.1
six==1.15.0
tqdm==4.51.0
Deprecated==1.2.10
wrapt==1.12.1
jsonfield==3.1.0

# Django Apps
django-allauth==0.43.0
django-appconf==1.0.4
django-filter==2.4.0
django-imagekit==4.0.2
django-jsonfield==1.4.0
django-jsonfield-compat==0.4.4
django-taggit==1.3.0
django-markdownify==0.8.1
django-mptt==0.11.0
django-modeltranslation>=0.11,<0.17.0
django-treebeard==4.3.1
django-guardian<2.3.1
django-downloadview==2.1.1
django-polymorphic==3.0.0
django-tastypie<0.15.0
django-tinymce==3.1.0
django-grappelli==2.14.2
oauthlib==3.1.0
pyopenssl==19.1.0

# geopython dependencies
pyproj==2.6.1.post1
OWSLib==0.20.0
pycsw==2.4.2
SQLAlchemy==1.3.20 # required by PyCSW
Shapely==1.7.1
mercantile==1.1.6
geoip2==4.1.0
numpy==1.19.*

# # Apps with packages provided in GeoNode's PPA on Launchpad.

# Django Apps
dj-database-url==0.5.0
dj-pagination==2.5.0
django-floppyforms<1.10.0
django-forms-bootstrap<=3.1.0
django-autocomplete-light==3.5.1
django-leaflet==0.27.1
django-invitations<1.9.4
django-recaptcha==2.0.6

# REST
djangorestframework>=3.1.0,<=3.12.1
djangorestframework-gis>=0.16
djangorestframework_guardian==0.3.0
drf-extensions==0.6.0
drf-writable-nested==0.6.1
drf-yasg==1.20.0
dynamic-rest==2.0.0
Markdown==3.3.3

pinax-notifications==6.0.0
pinax-ratings==4.0.0

# GeoNode org maintained apps.
django_mapstore_adapter==2.0.5.1
django-geonode-mapstore-client==2.0.8.1
geonode-avatar==5.0.7
django-geonode-client==1.0.9
geonode-oauth-toolkit==2.1.0
geonode-user-messages==2.0.2
geonode-announcements==2.0.2
geonode-django-activity-stream==0.9.0
gn-arcrest==10.5.3
geonode-dialogos==3.0.1
geoserver-restconfig==2.0.4.8
gn-gsimporter==2.0.1
gisdata==0.5.4

# haystack/elasticsearch
django-haystack==3.0
elasticsearch>=2.0.0,<8.0.0

# datetimepicker widget
django-bootstrap3-datetimepicker-2==2.8.2

# AWS S3 dependencies
django-storages==1.10.1

# Django Caches
python-memcached<=1.59

# Contribs
xmltodict<0.12.1
inflection==0.3.1
jdcal==1.4.1
mock<5.0.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
timeout-decorator==0.4.1
python_resize_image==1.1.19

# required by monitoring
psutil==5.7.3
django-cors-headers==3.5.0
user-agents
xmljson
django-ipware<3.1
# no version here, use latest one with fresh data
pycountry

# production
uWSGI==2.0.19.1
gunicorn==20.0.4
ipython==7.18.1
docker==4.3.1
invoke==1.4.1

# tests
coverage==5.3
parse-type==0.5.2
requests-toolbelt==0.9.1
flake8==3.8.4
pytest==6.1.1
pytest-bdd==4.0.1
splinter==0.14.0
pytest-splinter==3.2.0
pytest-django==4.1.0
setuptools==50.3.2
pip==20.2.4
Twisted==20.3.0
factory-boy==3.1.0
flaky==3.7.0
